Remove tests checking internals of DynamicILInfo
